Refactored the Archive.Today handler to use Selenium.
Archive.Today added a captcha test when making a POST request to submit a page to be archived. When you manually go to the website and submit a page, there is no captcha. Selenium makes it possible to submit a URL to be archived through the Archive.Today website without encountering a captcha test.